refactor:Modificacion de tipo de datos y vinculacion de BD

+++ backend/src/main/java/com/neon/sve/dto/detalleIngreso/DatosListadoDetalleIngreso.java
@@ -1,5 +1,6 @@
 package com.neon.sve.dto.detalleIngreso;
 
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 
 import com.neon.sve.model.stock.DetalleIngreso;
@@ -13,8 +14,8 @@ public record DatosListadoDetalleIngreso(
         String modelo,
         String codigoLote,
         int cantidad,
-        double precioUnitario,
-        double subtotal
+        BigDecimal precioUnitario,
+        BigDecimal subtotal
 
 ) {
